+++ src/InstrumentationEngine/MethodInfo.cpp
@@ -38,7 +38,6 @@ MicrosoftInstrumentationEngine::CMethodInfo::CMethodInfo(
     m_pSig(nullptr),
     m_cbSigBlob(0),
     m_dwILStreamLen(0),
-    m_dwCorILMapmLen(0),
     m_localVariables(nullptr),
     m_origLocalVariables(nullptr),
     m_cbIntermediateRenderedMethod(0),
@@ -1346,16 +1345,16 @@ HRESULT MicrosoftInstrumentationEngine::CMethodInfo::ApplyFinalInstrumentation(b
         CLogging::LogError(_T("CMethodInfo::ApplyFinalInstrumentation should only be called if a method body has been set for this function"));
         return E_FAIL;
     }
-    
+
+    CComPtr<CProfilerManager> pProfilerManager;
+    IfFailRet(CProfilerManager::GetProfilerManagerInstance(&pProfilerManager));
+
+    CComPtr<ICorProfilerInfo> pCorProfilerInfo;
+    IfFailRet(pProfilerManager->GetRealCorProfilerInfo(&pCorProfilerInfo));
+
     if (!isRejit)
     {
         CLogging::LogMessage(_T("CMethodInfo::ApplyFinalInstrumentation - Non-rejit case"));
-
-        CComPtr<CProfilerManager> pProfilerManager;
-        IfFailRet(CProfilerManager::GetProfilerManagerInstance(&pProfilerManager));
-
-        CComPtr<ICorProfilerInfo> pCorProfilerInfo;
-        IfFailRet(pProfilerManager->GetRealCorProfilerInfo(&pCorProfilerInfo));
 
         ModuleID moduleId;
         IfFailRet(m_pModuleInfo->GetModuleID(&moduleId));
@@ -1377,9 +1376,9 @@ HRESULT MicrosoftInstrumentationEngine::CMethodInfo::ApplyFinalInstrumentation(b
         m_pModuleInfo->SetMethodIsTransformed(m_tkFunction, true);
 
         // This will fail if no entries in CorILMap, plus there's no point in setting this in such a case anyway.
-        if (m_dwCorILMapmLen > 0)
+        if (m_pCorILMap.Count() > 0)
         {
-            IfFailRet(pCorProfilerInfo->SetILInstrumentedCodeMap(m_functionId, TRUE, m_dwCorILMapmLen, m_pCorILMap));
+            IfFailRet(pCorProfilerInfo->SetILInstrumentedCodeMap(m_functionId, TRUE, (ULONG)m_pCorILMap.Count(), m_pCorILMap.Get()));
         }
     }
     else
@@ -1413,16 +1412,11 @@ HRESULT MicrosoftInstrumentationEngine::CMethodInfo::ApplyFinalInstrumentation(b
         GetRejitCount(&rejitCount);
         if (rejitCount != 0)
         {
-            CAutoVectorPtr<COR_IL_MAP> pTempCorILMap;
-            DWORD corILMapmLen = m_dwCorILMapmLen;
+            DWORD corILMapmLen = (DWORD)m_pCorILMap.Count();
             corILMapmLen++;
-            pTempCorILMap.Allocate(corILMapmLen);
-            if (m_pCorILMap == NULL)
-            {
-                return E_OUTOFMEMORY;
-            }
+            SharedArray<COR_IL_MAP> pTempCorILMap(corILMapmLen);
 
-            memcpy(pTempCorILMap, m_pCorILMap, m_dwCorILMapmLen * sizeof(COR_IL_MAP));
+            memcpy(pTempCorILMap.Get(), m_pCorILMap.Get(), m_pCorILMap.Count() * sizeof(COR_IL_MAP));
 
             pTempCorILMap[corILMapmLen - 1].fAccurate = true;
             pTempCorILMap[corILMapmLen - 1].oldOffset = 0xfeefee;
@@ -1430,13 +1424,13 @@ HRESULT MicrosoftInstrumentationEngine::CMethodInfo::ApplyFinalInstrumentation(b
 
             // NOTE: CLR's earlier than 4.5.2 won't have this implemented and will return E_NOTIMPL.
             // 4.5.2 is a prereq for rejit.
-            m_pFunctionControl->SetILInstrumentedCodeMap(corILMapmLen, pTempCorILMap);
+            m_pFunctionControl->SetILInstrumentedCodeMap(corILMapmLen, pTempCorILMap.Get());
         }
         else
         {
             // NOTE: CLR's earlier than 4.5.2 won't have this implemented and will return E_NOTIMPL.
             // 4.5.2 is a prereq for rejit.
-            m_pFunctionControl->SetILInstrumentedCodeMap(m_dwCorILMapmLen, m_pCorILMap);
+            m_pFunctionControl->SetILInstrumentedCodeMap((ULONG)m_pCorILMap.Count(), m_pCorILMap.Get());
         }
     }
 
@@ -1927,72 +1921,68 @@ HRESULT MicrosoftInstrumentationEngine::CMethodInfo::MergeILInstrumentedCodeMap(
 
     CLogging::LogMessage(_T("Start CMethodInfo::SetILInstrumentedCodeMap"));
 
-    if (m_pCorILMap != NULL)
+    if (!m_pCorILMap.IsEmpty())
     {
-        // The raw callbacks will assume the updated il stream that came out of the instrumentation methods and/or
-        // other raw callbacks. The maps must be merged.
-        //
-        // Merge algorithm is as follows:
-        // 1) For each entry in the new map, look up the original offset in the old map's new offsets
-        // 2) If a match is found, replace the old offset in the new map with the original offset from the old map.
-        // 3) If no match is found, then this offset did not exist in the original map and its entry should be removed.
-
-        // First index the old entries by new offset.
-        std::unordered_map<ULONG32, COR_IL_MAP*> oldMapNewOffsetToEntryMap;
-        for (DWORD i = 0; i < m_dwCorILMapmLen; i++)
+        // No need to merge if we are just setting to the exact same pointer.
+        if (rgILMapEntries != m_pCorILMap.Get())
         {
-            COR_IL_MAP* pCurr =  &(m_pCorILMap[i]);
-            oldMapNewOffsetToEntryMap[pCurr->newOffset] = pCurr;
-        }
+            // The raw callbacks will assume the updated il stream that came out of the instrumentation methods and/or
+            // other raw callbacks. The maps must be merged.
+            //
+            // Merge algorithm is as follows:
+            // 1) For each entry in the new map, look up the original offset in the old map's new offsets
+            // 2) If a match is found, replace the old offset in the new map with the original offset from the old map.
+            // 3) If no match is found, then this offset did not exist in the original map and its entry should be removed.
 
-        // Next, iterate over the new map doing the conversion.
-        std::vector<COR_IL_MAP*> updatedEntries;
-        for (DWORD k = 0; k < cILMapEntries; k++)
-        {
-             COR_IL_MAP* pCurr = &(rgILMapEntries[k]);
+            // First index the old entries by new offset.
+            std::unordered_map<ULONG32, COR_IL_MAP*> oldMapNewOffsetToEntryMap;
+            size_t currEntries = m_pCorILMap.Count();
+            for (DWORD i = 0; i < currEntries; i++)
+            {
+                COR_IL_MAP* pCurr = &(m_pCorILMap[i]);
+                oldMapNewOffsetToEntryMap[pCurr->newOffset] = pCurr;
+            }
 
-             std::unordered_map<ULONG32, COR_IL_MAP*>::iterator iter = oldMapNewOffsetToEntryMap.find(pCurr->oldOffset);
+            // Next, iterate over the new map doing the conversion.
+            std::vector<COR_IL_MAP*> updatedEntries;
+            for (DWORD k = 0; k < cILMapEntries; k++)
+            {
+                COR_IL_MAP* pCurr = &(rgILMapEntries[k]);
 
-             if (iter != oldMapNewOffsetToEntryMap.end())
-             {
-                 COR_IL_MAP* pOldEntry = iter->second;
+                std::unordered_map<ULONG32, COR_IL_MAP*>::iterator iter = oldMapNewOffsetToEntryMap.find(pCurr->oldOffset);
 
-                 // Entry was found in the original map.
-                 pCurr->oldOffset = pOldEntry->oldOffset;
+                if (iter != oldMapNewOffsetToEntryMap.end())
+                {
+                    COR_IL_MAP* pOldEntry = iter->second;
 
-                 updatedEntries.push_back(pCurr);
-             }
-             else
-             {
-                 // No entry was found. Skip this entry. It must have been a new instruction in the old map.
-             }
-        }
+                    // Entry was found in the original map.
+                    pCurr->oldOffset = pOldEntry->oldOffset;
 
-        m_pCorILMap.Free();
-        m_dwCorILMapmLen = (DWORD)updatedEntries.size();
-        m_pCorILMap.Allocate(m_dwCorILMapmLen);
-        if (m_pCorILMap == NULL)
-        {
-            return E_OUTOFMEMORY;
-        }
+                    updatedEntries.push_back(pCurr);
+                }
+                else
+                {
+                    // No entry was found. Skip this entry. It must have been a new instruction in the old map.
+                }
+            }
 
-        for (DWORD x = 0; x < m_dwCorILMapmLen; x++)
-        {
-            m_pCorILMap[x] = *(updatedEntries[x]);
+            m_pCorILMap.Reset(SharedArray<COR_IL_MAP>(updatedEntries.size()));
+
+            for (DWORD x = 0; x < m_pCorILMap.Count(); x++)
+            {
+                m_pCorILMap[x] = *(updatedEntries[x]);
+            }
         }
     }
     else
     {
-        m_dwCorILMapmLen = cILMapEntries;
-        m_pCorILMap.Allocate(m_dwCorILMapmLen);
-        if (m_pCorILMap == NULL)
-        {
-            return E_OUTOFMEMORY;
-        }
+        m_pCorILMap.Reset(SharedArray<COR_IL_MAP>(cILMapEntries));
 
-        memcpy(m_pCorILMap, rgILMapEntries, m_dwCorILMapmLen * sizeof(COR_IL_MAP));
+        memcpy(m_pCorILMap.Get(), rgILMapEntries, cILMapEntries * sizeof(COR_IL_MAP));
     }
 
+    // Save the map with the module info so that it can be retrieved later by the JIT callbacks.
+    m_pModuleInfo->SetILInstrumentationMap(m_tkFunction, m_pCorILMap);
     CLogging::LogMessage(_T("End CMethodInfo::SetILInstrumentedCodeMap"));
 
     return S_OK;
@@ -2049,9 +2039,9 @@ HRESULT MicrosoftInstrumentationEngine::CMethodInfo::GetInstrumentationResults(
     *ppInstructionGraph = pInstructionGraph.Detach();
     *ppExceptionSection = pExceptionSection.Detach();
 
-    *ppCorILMap = new COR_IL_MAP[m_dwCorILMapmLen];
-    memcpy(*ppCorILMap, m_pCorILMap, m_dwCorILMapmLen * sizeof(COR_IL_MAP));
-    *dwCorILMapmLen = m_dwCorILMapmLen;
+    *ppCorILMap = new COR_IL_MAP[m_pCorILMap.Count()];
+    memcpy(*ppCorILMap, m_pCorILMap.Get(), m_pCorILMap.Count() * sizeof(COR_IL_MAP));
+    *dwCorILMapmLen = (DWORD)m_pCorILMap.Count();
 
     return hr;
 }